use jobs class
